requirements.txt
aiogram>=3.0.0
sqlalchemy
aiosqlite
apscheduler
pandas
openpyxl
python-dotenv
asyncpg
psycopg2-binary
pytz